feat: catch URI parse errors when converting PSR-7 request to laminas-http
This patch adds a try/catch block within the `LaminasRouter::match()` method around the conversion of the PSR-7 request to a laminas-http request.
It catches laminas-http `InvalidArgumentException`s, and checks for a previous exception matching a laminas-http `InvalidUriPartException`.
When detected, it returns a `RouteResult` indicating a routing failure; otherwise, it rethrows the exception.

The patch is intended to fix scenarios where the Host header is invalid (e.g. `${ip}`), leading to creation of an invalid URI.

Fixes #12

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>